consolidation into one primary caching method

+++ wpcachedb.class.php
@@ -1,10 +1,11 @@
 <?php
 /**
  * Simple WordPress Database Caching for Developers
+ *
  * @package WordPress
  * @since 3.0
  *
- * instructions comming soon ...
+ * ... instructions comming soon.
  *
  */
 define(CACHE_DIR,dirname(__FILE__).'/cache');
@@ -31,14 +32,17 @@ class WPCacheDB extends wpdb {
 	 *
 	 * More information can be found on the codex page.
 	 *
-	 * @since 0.71
-	 *
+	 * @since 0.1
+	 *	
+	 * @param string $method WPDB method to use
 	 * @param string $query Database query
-	 * @return int|false Number of rows affected/selected or false on error
+	 * @param string $output WPDB output option ( OBJECT, ARRAY_A, ARRAY_N, etc... )
+	 * @param string $flag Used for creating cache groups
+	 * @param int number of seconds to store the cache. Will override class default 
+	 * @return object/array of data rerieve via the wpdb method.
 	 */
-		
 	
-	function cached_result($query,$output,$flag ='',$time = '') {
+	function cache($method, $query,$output,$flag ='',$time = '') {
 		if($time != '') {
 			$this->cache_time = $time; 
 		} 
@@ -75,7 +79,12 @@ class WPCacheDB extends wpdb {
 					return $result;
 					exit();	
 		} else {
-				$result = $this->get_results($query,$output);
+				if('get_var'==$method)
+				{
+					$result = call_user_func_array(array($this,$method),array($query));	
+				} else {
+					$result = call_user_func_array(array($this,$method),array($query,$output));
+				}
 				if(!empty($result)) {
 					file_put_contents($this->cache_file_name,serialize($result)); 
 					return $result;
@@ -85,20 +94,32 @@ class WPCacheDB extends wpdb {
 		file_put_contents($this->cache_log_file,serialize($this->cache_log));
 	}
 	
+	/**
+	 * Clears the cache
+	 *
+	 * @since 0.1
+	 * @param string $flag Flag to clear
+	 */
+	
 	public static function clear($flag = '') {
 		if($flag != '') {
 			$flag = trim($flag,'/').'/';
 		} 
 		
 		$dir = CACHE_DIR.'/'.$flag;
+		if(is_dir($dir)) {
 		$mydir = opendir(CACHE_DIR.'/'.$flag);
-    while(false !== ($file = readdir($mydir))) {
-        if($file != "." && $file != "..") {
-            chmod($dir.$file, 0777);
-           	unlink($dir.$file) or DIE("couldn't delete $dir$file<br />");
-        }
-    }
-    closedir($mydir);
+	    while(false !== ($file = readdir($mydir))) {
+	        if($file != "." && $file != "..") {
+	            chmod($dir.$file, 0777);
+	           	unlink($dir.$file) or DIE("couldn't delete $dir$file<br />");
+	        }
+	    }
+	    closedir($mydir);
+	  }
 	}
 	
+	
+	
 }
+?>
